fix: merge length util with number and string
