+++ packages/utility/src/number/index.ts
@@ -1,4 +1,3 @@
-export * from './length';
 export * from './percentage';
 export * from './round';
 export * from './sum';
