--- packages/utility/src/number/length.ts
@@ -1,12 +1,0 @@
-export const length = (
-  val: number | string | null | undefined,
-  isIncludePoint = false,
-) => {
-  let str = typeof val === 'number' ? val.toString() : val || '';
-
-  if (!isIncludePoint) {
-    str = str.replace('.', '');
-  }
-
-  return str.length;
-};
