+++ packages/utility/src/object/index.ts
@@ -1,2 +1,3 @@
 export * from './findKey';
+export * from './length';
 export * from './hasKeys';
