+++ packages/utility/src/string/index.ts
@@ -1,2 +1,1 @@
-export * from './length';
 export * from './toUpperSnakeCase';
