--- packages/utility/src/string/length.ts
@@ -1,6 +1,0 @@
-import GraphemeSplitter from 'grapheme-splitter';
-
-export const length = (str: string): number => {
-  const graphemeSplitter = new GraphemeSplitter();
-  return graphemeSplitter.countGraphemes(str);
-};
